Finished writing test function
Added clarification for add_vendor route

Ignore .vscode
